requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
sqlparse==0.5.1
clarifai==10.10.0
azure-storage-blob==12.24.0

pandas~=2.2.3
azure-core~=1.32.0
RapidFuzz~=3.10.1
gunicorn==22.0.0


requests~=2.32.3
pillow~=11.0.0